requirements.txt
streamlit
torch
numpy
joblib
pymongo
bcrypt
python-dotenv